+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.11.0] - 2026-09-08
+
 ### Fixed
 
 - `.github/workflows/release.yml`'s `verify` job now checks out full git history